refactor(kernel): move mha_decode_scheduler_metadata into FA3 wrapper
Define mha_decode_scheduler_metadata inside
tokenspeed_kernel/ops/attention/flash_attn/__init__.py beside the FA3
kernel registrations it wraps, and re-export it from
tokenspeed_kernel/ops/attention. The wrapper is the natural home for an
FA3-specific helper, per review comment on #81.

No behavior change: the top-level tokenspeed_kernel re-export and the
MHA backend caller are unchanged.
