+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/attention/flash_attn/__init__.py
@@ -27,6 +27,14 @@ from tokenspeed_kernel.platform import (
 )
 from tokenspeed_kernel.registry import Priority, error_fn, register_kernel
 
+__all__ = [
+    "flash_attn_func",
+    "flash_attn_varlen_func",
+    "flash_attn_with_kvcache",
+    "get_scheduler_metadata",
+    "mha_decode_scheduler_metadata",
+]
+
 flash_attn_func = error_fn
 flash_attn_varlen_func = error_fn
 flash_attn_with_kvcache = error_fn
@@ -401,9 +409,36 @@ elif platform.is_nvidia and platform.is_hopper:
 # ------------------------------------------------------------------------------
 
 
-__all__ = [
-    "flash_attn_func",
-    "flash_attn_varlen_func",
-    "flash_attn_with_kvcache",
-    "get_scheduler_metadata",
-]
+def mha_decode_scheduler_metadata(
+    *,
+    batch_size: int,
+    max_seqlen_q: int,
+    max_seqlen_k: int,
+    num_heads_q: int,
+    num_heads_kv: int,
+    headdim: int,
+    cache_seqlens: torch.Tensor,
+    qkv_dtype: torch.dtype,
+    page_size: int,
+    causal: bool = True,
+) -> torch.Tensor | None:
+    """Pre-compute decode scheduler metadata once per scheduler step.
+
+    Only the FA3 decode kernel consumes pre-computed scheduler metadata; on
+    every other backend the kernel computes it internally and this helper
+    returns ``None`` so callers can pass through unconditionally.
+    """
+    if get_scheduler_metadata is error_fn:
+        return None
+    return get_scheduler_metadata(
+        batch_size=batch_size,
+        max_seqlen_q=max_seqlen_q,
+        max_seqlen_k=max_seqlen_k,
+        num_heads_q=num_heads_q,
+        num_heads_kv=num_heads_kv,
+        headdim=headdim,
+        cache_seqlens=cache_seqlens,
+        qkv_dtype=qkv_dtype,
+        page_size=page_size,
+        causal=causal,
+    )
